requirements.txt
typing_extensions
paho-mqtt
websocket-client
websockets
kodi-json
